move the plugin classes to io.github.flibiostudio.updatifier package
