+++ src/main/java/io/github/flibiostudio/updatifier/FileManager.java
@@ -23,7 +23,7 @@
  * THE SOFTWARE.
  */
 
-package me.flibio.updatifier;
+package io.github.flibiostudio.updatifier;
 
 import com.google.common.reflect.TypeToken;
 import ninja.leaping.configurate.ConfigurationNode;
